KEGGAPI.jl Functions

KEGGAPI.conv — Method

KEGGAPI.conv(<targetdb>, <sourcedb>)

Convert KEGG identifiers to/from outside identifiers.

Example:

using KEGGAPI
KEGGAPI.conv("eco", "ncbi-geneid")
KEGGAPI.conv("ncbi-geneid", "eco")
KEGGAPI.conv("ncbi-proteinid", "hsa:10458+ece:Z5100")
KEGGAPI.conv("genes", "ncbi-geneid:948364")

KEGGAPI.find — Function

KEGGAPI.find(database, query) -> KeggTupleList

Find entries in a specific database from the KEGG API.

using KEGGAPI
KEGGAPI.find("compound","glucose")

KEGGAPI.get_image — Method

KEGGAPI.get_image(pathway) -> Image

Get an image of a specific pathway from the KEGG API.

Example:

KEGGAPI.get_image("hsa00010")

KEGGAPI.info — Method

KEGGAPI.info(database) -> String

Get information about a specific database from the KEGG API.

Example:

KEGGAPI.info("kegg")

KEGGAPI.kegg_get — Function

KEGGAPI.get(query, option) -> Vector

Get entries from across the KEGG database. The get function takes a vector of strings and returns a vector containing the urls used and the data returned.

The get function will split the vector into groups of 10 and make a request for each chunk.

Examples

using KEGGAPI
kegg_get_genes = 
    KEGGAPI.kegg_get(
        ["hsa:10458", "hsa:10458", "hsa:10458", "hsa:10458"], 
        "ntseq"
        );
first(kegg_get_genes)
kegg_get_compounds = 
KEGGAPI.kegg_get(
    ["C01290","G00092"]
    );
first(kegg_get_compounds)

KEGGAPI.link — Method

KEGGAPI.conv(<targetdb>, <sourcedb>)

Find related entries by using database cross-references

Example:

KEGGAPI.link("pathway", "hsa")

KEGGAPI.list — Function

KEGGAPI.list(database)

Get a list of entries from a specific database from the KEGG API.

Example:

KEGGAPI.list("pathway")

KEGGAPI.request — Method

request(url)

Make a request to the specified URL and return the response body as a string. If an error occurs, a RequestError is thrown. This function is not intended for direct use. Instead, use the info, list, and find, etc. functions.

Examples

request("https://rest.kegg.jp/info/kegg")

KEGGAPI.request_other — Method

request_other(url)

Make a request to the specified URL and return the response body as a vector. If an error occurs, a RequestError is thrown. This function is not intended for direct use.

Examples

request_other("https://rest.kegg.jp/image/hsa00010")

KEGGAPI.save_image — Method

KEGGAPI.save_image(image, filename) -> filename

Save an image to a file.

Examples

image = KEGGAPI.get_image("hsa00010")
KEGGAPI.save_image(image, "glycolysis.png")