KEGGAPI

Documentation for KEGGAPI.

KEGGAPI.get_image

KEGGAPI.request

KEGGAPI.get_image — Method

KEGGAPI.get_image(pathway) -> Image

Get an image of a specific pathway from the KEGG API.

Examples

julia> KEGGAPI.get_image("hsa00010")

KEGGAPI.request — Method

request(url)

Make a request to the specified URL and return the response body as a string. If an error occurs, a RequestError is thrown. This function is not intended for direct use. Instead, use the info, list, and find, etc. functions.

Examples

julia> request("https://rest.kegg.jp/info/kegg")